[RSDK-8124] Extract serve_async_with_external_network to common
